refactor: friend-schedule 베럴패턴 적용
