--- src/features/friend-schedule/index.ts
@@ -1,7 +1,0 @@
-// API
-export * from './api/friend-schedule.API'
-
-// UI
-export { default as ShareSchedule } from './ui/friend-calendar'
-export { default as FriendDetailedScheduleList } from './detailed-schedule/friend-detailed-schedule-list'
-export { default as FriendDetailedSchedule } from './detailed-schedule/friend-detailed-schedule'
