+++ src/features/friend-schedule/ui/friend-detailed-schedule-list.tsx
