+++ src/features/friend-schedule/ui/friend-detailed-schedule.tsx
