+++ src/features/friend-schedule/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as FriendCalendar } from './friend-calendar'
+export { default as FriendDetailedScheduleList } from './friend-detailed-schedule-list'
+export { default as FriendDetailedSchedule } from './friend-detailed-schedule'
