+++ src/pages/detailed-schedule/friend-detailed-schedule/friend-detailed-schedule-list-page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react'
 
-import FriendDetailedScheduleList from '@/features/friend-schedule/detailed-schedule/friend-detailed-schedule-list'
+import FriendDetailedScheduleList from '@/features/friend-schedule/ui/friend-detailed-schedule-list'
 
 import DetailedScheduleListSkeleton from '../detailed-schedule-list-skeleton'
